Added exclusion also to the partial preload
