+++ inc/Engine/Preload/Controller/CheckExcludedTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace WP_Rocket\Engine\Preload\Controller;
+
+trait CheckExcludedTrait
+{
+	/**
+	 * Check if the url is excluded.
+	 *
+	 * @param string $url url to check.
+	 * @return bool
+	 */
+	protected function check_excluded( string $url ) {
+		$excluded = get_rocket_cache_reject_uri();
+		return (bool) preg_match( "/$excluded/", $url );
+	}
+}
