+++ tests/Unit/inc/Engine/Preload/Controller/ClearCache/partialClean.php
@@ -2,6 +2,7 @@
 
 namespace WP_Rocket\Tests\Unit\inc\Engine\Preload\Controller\ClearCache;
 
+use Mockery;
 use WP_Rocket\Engine\Preload\Controller\ClearCache;
 use WP_Rocket\Engine\Preload\Database\Queries\Cache;
 use WP_Rocket\Tests\Unit\TestCase;
@@ -19,14 +20,20 @@ class Test_PartialClean extends TestCase
 	{
 		parent::setUp();
 		$this->query = $this->createMock(Cache::class);
-		$this->controller = new ClearCache($this->query);
+		$this->controller = Mockery::mock(ClearCache::class . '[check_excluded]', [$this->query])->shouldAllowMockingProtectedMethods();
 	}
 
 	/**
 	 * @dataProvider configTestData
 	 */
 	public function testShouldDoAsExpected($config, $expected) {
-		$this->query->expects(self::atLeastOnce())->method('create_or_update')->withConsecutive(...$expected['urls']);
+
+		foreach ($config['urls'] as $url) {
+			$this->controller->expects()->check_excluded($url)->andReturn($config['is_excluded']);
+		}
+		if(! $config['is_excluded']) {
+			$this->query->expects(self::atLeastOnce())->method('create_or_update')->withConsecutive(...$expected['urls']);
+		}
 
 		$this->controller->partial_clean($config['urls']);
 	}
